adding ngpopups and minor fixes

+++ src/app/components/user/profile/profile.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { NgPopupsService } from 'ng-popups';
 import { ToastrService } from 'ngx-toastr';
 import { User } from 'src/app/models/User';
 import { UserService } from 'src/app/services/user.service';
@@ -14,6 +15,7 @@ export class ProfileComponent implements OnInit {
   path: string;
 
   constructor(
+    private popups: NgPopupsService,
     private router: Router,
     private toast: ToastrService,
     private userService: UserService
@@ -29,8 +31,21 @@ export class ProfileComponent implements OnInit {
   }
 
   deleteUser() {
-    console.warn('really delete?');
-    this.userService.delete();
+    this.popups
+      .confirm(
+        'Do You really want to delete your Account?',
+        {
+          title: 'Think twice!',
+          okButtonText: 'Yes, delete me!',
+          cancelButtonText: 'No',
+          color: '#F00'
+        }
+      )
+      .subscribe(res => {
+        if (res) {
+          this.userService.delete();
+        }
+      });
   }
 
   resendVerificationEmail() {
